% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFATools.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{Plotting a confusion matrix heatmap}
\usage{
plot_confusion_matrix(confusion_matrix, title = "Confusion Matrix Heatmap")
}
\arguments{
\item{confusion_matrix}{a caret confusionMatrix object created using caret::confusionMatrix()}

\item{title}{the plot title}
}
\value{
A ggplot object
}
\description{
Creates a heatmap visualization from a caret confusionMatrix object.
The diagonal cells represent correct classifications, while off-diagonal
cells represent misclassifications
}
\examples{
# Example using iris dataset
cm <- caret::confusionMatrix(
  data = iris$Species,
  reference = iris$Species
)

plot_confusion_matrix(cm, title = "Example Confusion Matrix")
}
\keyword{confusion-matrix}
\keyword{heatmap}
\keyword{visualization}
