% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFATools.R
\name{prepare_data}
\alias{prepare_data}
\title{Preparing data for Discriminant Function Analysis}
\usage{
prepare_data(
  data,
  target_column_name,
  training_propotion_size,
  scale = TRUE,
  seed = 123
)
}
\arguments{
\item{data}{a data frame}

\item{target_column_name}{A character string stating the categorical response variable name}

\item{training_propotion_size}{The proportion of data used for training.}

\item{scale}{A logical value (TRUE or FALSE); if TRUE, the numeric predictors are centered and scaled}

\item{seed}{A random seed for reproducibility of results}
}
\value{
A list containing:
\item{training_data}{The processed training data.}
\item{test_data}{The processed test data.}
\item{preprocess_object}{The preprocessing object used for scaling. NULL if scale = FALSE.}
}
\description{
This function splits a dataset into training and testing data, and also optionally standardizes the predictor variables, when the scaling is used
,the centering and the scaling parameters are estimated from the training data and then aopplied to both training and test data after splitting to avoid data leakage.
}
\examples{
prepare_data(
  data = iris, target_column_name = "Species",
  training_propotion_size = 0.8
)
prepare_data(
  data = iris, target_column_name = "Species",
  training_propotion_size = 0.8, scale = FALSE
)
}
\keyword{discriminant-function-analysis}
