% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFATools.R
\name{run_fda_model}
\alias{run_fda_model}
\title{Running a Flexible Discriminant Analysis (FDA)}
\usage{
run_fda_model(training_data, test_data, target_column_name)
}
\arguments{
\item{training_data}{a data frame containing the training data}

\item{test_data}{a data frame containing the test data}

\item{target_column_name}{a character string of the categorical response variable name}
}
\value{
A list containing:
\item{model}{The fitted FDA model.}
\item{predictions}{The predicted class labels.}
\item{confusion_matrix}{A caret confusion matrix object.}
\item{accuracy}{The test set accuracy.}
}
\description{
#' Fits a Flexible Discriminant Analysis model and evaluates its classification
performance on test data. FDA is a more flexible extension of discriminant
analysis that can model more complex relationships between predictors and
class membership
}
\examples{
prepared <- prepare_data(iris, "Species", training_propotion_size=0.7)
fda_result <- run_fda_model(
  training_data = prepared$training_data,
  test_data = prepared$test_data,
  target_column_name = "Species"
)
fda_result$accuracy
}
\keyword{fda}
