% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFATools.R
\docType{data}
\name{spotify_data}
\alias{spotify_data}
\title{Spotify Dataset}
\format{
A data frame with multiple variables:
\describe{
\item{acousticness}{Numeric (0–1)}
\item{album_name}{Character}
\item{artists}{Character}
\item{danceability}{Numeric (0–1)}
\item{duration_ms}{Numeric}
\item{energy}{Numeric (0–1)}
\item{explicit}{Logical}
\item{instrumentalness}{Numeric (0–1)}
\item{key}{Integer}
\item{liveness}{Numeric}
\item{loudness}{Numeric (dB)}
\item{mode}{Integer}
\item{popularity}{Integer (0–100)}
\item{speechiness}{Numeric (0–1)}
\item{tempo}{Numeric}
\item{time_signature}{Integer}
\item{track_genre}{Character}
\item{track_id}{Character}
\item{track_name}{Character}
\item{valence}{Numeric}
}
}
\usage{
spotify_data
}
\description{
A dataset containing Spotify track features used for discriminant analysis.
}
\keyword{datasets}
